Change the test syntax
The "will successfully beat the random AI 25 times" test has been renamed the "will successfully beat the Medium AI 25 times" to point out that the AI that I am using to test the game AI is not entirely random. It does have some intelligence.

I also replaced the deprecated syntax of 'stub' in favor of the more modern syntax of "allow(object).to receive(:method)".